[refactor] facade 네이밍 변경
