+++ src/main/java/com/jinddung2/givemeticon/domain/trade/facade/TradeWriteFacade.java
@@ -6,7 +6,6 @@ import com.jinddung2.givemeticon.domain.notification.domain.dto.CreateNotificati
 import com.jinddung2.givemeticon.domain.notification.producer.NotificationProducer;
 import com.jinddung2.givemeticon.domain.sale.domain.Sale;
 import com.jinddung2.givemeticon.domain.sale.service.SaleService;
-import com.jinddung2.givemeticon.domain.trade.controller.dto.TradeDto;
 import com.jinddung2.givemeticon.domain.trade.domain.Trade;
 import com.jinddung2.givemeticon.domain.trade.exception.AlreadyBoughtSaleException;
 import com.jinddung2.givemeticon.domain.trade.service.TradeService;
@@ -14,13 +13,9 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.time.LocalDate;
-import java.time.temporal.ChronoUnit;
-import java.util.List;
-
 @Service
 @RequiredArgsConstructor
-public class TradeSaleItemUserFacade {
+public class TradeWriteFacade {
 
     private final SaleService saleService;
     private final ItemService itemService;
@@ -36,40 +31,13 @@ public class TradeSaleItemUserFacade {
         }
 
         Item item = itemService.getItem(sale.getItemId());
-        long restDay = getRestDay(sale.getExpirationDate());
-        Trade trade = Trade.builder()
-                .buyerId(buyerId)
-                .saleId(saleId)
-                .isUsed(false)
-                .build();
+        int tradeId = tradeService.save(sale, item, buyerId);
 
         sale.updateBoughtState();
         saleService.update(sale);
         producer.create(new CreateNotificationRequestDto(saleId, sale.getSellerId(),
                 String.format("%s이(가) 판매되었습니다.", item.getName())));
-        return tradeService.save(trade, item, restDay);
-    }
-
-    public TradeDto getTradeDetail(int tradeId, int buyerId) {
-        Trade trade = tradeService.getTrade(tradeId);
-        Sale sale = saleService.getSale(trade.getSaleId());
-        Item item = itemService.getItem(sale.getItemId());
-
-        return TradeDto.of(trade, sale, item);
-    }
-
-    public List<TradeDto> getUnusedTradeHistory(int buyerId, boolean orderByBoughtDate,
-                                                boolean orderByExpiredDate, int page) {
-
-        List<Trade> myUnusedItemHistory = tradeService.getMyUnusedItemHistory(buyerId, orderByBoughtDate, orderByExpiredDate, page);
-
-        return myUnusedItemHistory.stream()
-                .map(trade -> {
-                    Sale sale = saleService.getSale(trade.getSaleId());
-                    Item item = itemService.getItem(sale.getItemId());
-                    return TradeDto.of(trade, sale, item);
-                })
-                .toList();
+        return tradeId;
     }
 
     public void buyConfirmation(int tradeId, int buyerId) {
@@ -79,9 +47,5 @@ public class TradeSaleItemUserFacade {
         Item item = itemService.getItem(sale.getItemId());
         producer.create(new CreateNotificationRequestDto(sale.getId(), sale.getSellerId(),
                 String.format("%s이(가) 구매 확정되었습니다.", item.getName())));
-    }
-
-    private long getRestDay(LocalDate expiredDate) {
-        return ChronoUnit.DAYS.between(LocalDate.now(), expiredDate);
     }
 }
